[MODIFY] diseño en paralelo con cards y addresses

+++ src/views/profile/ProfilePaymentView.vue
@@ -4,21 +4,28 @@
   <v-container>
     <p class="text-h4 font-weight-medium mb-2">Métodos de pago</p>
 
-    <v-row v-if="cards.length > 0" v-for="card in cards" :key="card.id">
-      <payment-method-component
-        :id="card.id"
-        :user="card.owner"
-        :last_four="card.last_four"
-        :onDeleteHandle="cardDelete"
+    <v-expand-transition>
+      <v-row v-if="!loading">
+        <payment-method-component
+          v-for="card in cards"
+          :key="card.id"
+          :id="card.id"
+          :user="card.owner"
+          :last_four="card.last_four"
+          :onDeleteHandle="cardDelete"
+        />
+      </v-row>
+    </v-expand-transition>
+
+    <v-row v-if="!loading && cards.length === 0">
+      <orders-not-found-component
+        icon="mdi-credit-card-off-outline"
+        advise="No tienes métodos de pago registradas"
+        recomendation="Agrega un método de pago para poder realizar tus compras"
+        disableActions
       />
     </v-row>
-    <v-row v-else>
-      <v-col cols="12">
-        <v-card variant="flat" class="text-center">
-          <v-card-title class="text-h6">No tienes tarjetas registradas</v-card-title>
-        </v-card>
-      </v-col>
-    </v-row>
+
     <v-btn
       variant="flat"
       class="mb-8 mt-10 text-none"
@@ -27,7 +34,7 @@
       prepend-icon="mdi-plus"
       :to="{ name: 'ProfileAddPayment' }"
     >
-      Agregar nueva tarjeta
+      Agregar método de pago
     </v-btn>
   </v-container>
 </template>
@@ -38,59 +45,68 @@ import { useProfileStore } from "@/store/ProfileStore";
 import { ref, onMounted, reactive } from "vue";
 import { Toast } from "@/utils/Alerts";
 import { decryptValue, encryptAES } from "@/utils/Crypto";
+import Swal from "sweetalert2";
 
 const { fetchCards, deleteCard } = useProfileStore();
 
 const cards = reactive([]);
-const loading = ref(false);
+const loading = ref(true);
 
 const getCards = async () => {
   try {
     const response = await fetchCards();
-    response.forEach(card => {
+    cards.splice(0);
+    response.forEach((card) => {
       cards.push({
         id: card.id,
         owner: decryptValue(card.owner),
         last_four: card.lastFour,
       });
-    })
-    loading.value = false;
+    });
   } catch (error) {
     Toast.fire({
       icon: "error",
       title: "Ocurrió un error al cargar tus tarjetas",
     });
-    loading.value = false;
   } finally {
     loading.value = false;
   }
 };
 
 const cardDelete = async (id) => {
-  loading.value = true;
   try {
-    const response = await deleteCard(encryptAES(String(id)));
-    window.location.reload();
-    if(response.status === 200) {
-      Toast.fire({
-        icon: "success",
-        title: "Tarjeta eliminada",
-      });
-      loading.value = false;
-      getCards();
+    const result = await Swal.fire({
+      title: "¿Estas seguro?",
+      text: "Este método de pago será eliminado permanentemente",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: colors.primary,
+      cancelButtonColor: colors.primary_dark,
+      confirmButtonText: "Si",
+      cancelButtonText: "No",
+    });
+
+    if (result.isConfirmed) {
+      const response = await deleteCard(encryptAES(String(id)));
+      if (response.status === 200) {
+        Toast.fire({
+          icon: "success",
+          title: "Tarjeta eliminada correctamente",
+        });
+      }
     }
   } catch (error) {
-    loading.value = false;
     Toast.fire({
       icon: "error",
       title: "Ocurrió un error al eliminar la tarjeta",
     });
   } finally {
-    loading.value = false;
+    getCards();
   }
 };
 
 onMounted(() => {
+  loading.value = true;
   getCards();
 });
